refactor: Migrate `hasura-client.js` to TS
